rename: navbar, sidebar directory구조 변경 #18
